refactor and allowing source file name to be specified
