+++ openmc_source_plotter/__init__.py
@@ -1,5 +1,5 @@
+from .utils import get_particle_data
+from .utils import create_initial_particles
 from .core import plot_direction_from_initial_source
 from .core import plot_position_from_initial_source
 from .core import plot_energy_from_initial_source
-from .core import get_particle_data
-from .core import create_initial_particles
